+++ Runtime/EditorComment.cs
@@ -22,85 +22,20 @@ namespace OcUtility
         
         public EditorCommentAsset asset;
 
-        public bool hideGizmo = true;
-        [HideIf(nameof(hideGizmo)), Indent()]public Color gizmoColor = Color.red;
-        [HideIf(nameof(hideGizmo)), Indent()]public bool blinkGizmo;
+        const string FolderPath = "Assets/Editor Default Resources/Editor Comments";
 
-        [ShowInInspector][ShowIf(nameof(_renameAsset)), DelayedProperty]
-        public string nameInputField
-        {
-            get => asset == null ? name : asset.name;
-            set
-            {
-                AssetDatabase.RenameAsset(AssetDatabase.GetAssetPath(asset), GetValidName(value));
-                name = value;
-                _renameAsset = false;
-            }
-        }
-        bool _renameAsset;
-
-        [ShowIf("@asset == null")][Button]
+        [DisableInPrefabs][ShowIf("@asset == null")][Button]
         void CreateAsset()
         {
-            var assetName = name == "GameObject" ? "New Comment" : name;
-            
-            var baseFolder = "Assets/Editor Default Resources/Editor Comments";
-            var folderPath = $"{baseFolder}/{gameObject.scene.name}";
-            
-            asset = CreateAsset(folderPath, assetName);
-
-            if (asset.name != "GameObject")
-            {
-                name = asset.name;
-            }
-        }
-        [Button, ShowIf("@asset != null && !_renameAsset")][HideIf(nameof(_renameAsset))]
-        void Rename()
-        {
-            _renameAsset = true;
+            asset = EditorCommentAsset.CreateAsset($"{FolderPath}/{gameObject.scene.name}", name);
         }
         
         void OnDrawGizmos()
         {
-            var color = blinkGizmo ? gizmoColor.SetA(Time.realtimeSinceStartup % 0.5f) : gizmoColor; 
-            Gizmos.color = color;
+            Gizmos.color = ColorExtension.Rainbow(5f).SetA(0.54f);
 
             var dist = Vector3.Distance(SceneView.lastActiveSceneView.camera.transform.position, transform.position);
             Gizmos.DrawSphere(transform.position, dist * 0.01f);
-        }
-
-
-        // static-------
-        static string GetValidName(string targetName)
-        {
-            for (int i = 0; true; i++)
-            {
-                var nextName = i == 0 ? targetName : $"{targetName} {i}";
-                var exist = AssetDatabase.FindAssets($"t:EditorCommentAsset")
-                    .Select(AssetDatabase.GUIDToAssetPath).Select(AssetDatabase.LoadAssetAtPath<EditorCommentAsset>);
-
-                if (exist.All(x => x.name != nextName))
-                {
-                    return nextName;
-                }
-            }
-        }
-        
-        public static EditorCommentAsset CreateAsset(string folderPath, string fileName)
-        {
-            var comment = ScriptableObject.CreateInstance<EditorCommentAsset>();
-            comment.name = GetValidName(fileName);
-
-            
-            if (!AssetDatabase.IsValidFolder(folderPath))
-            {
-                System.IO.Directory.CreateDirectory(folderPath);
-                AssetDatabase.ImportAsset(folderPath);
-            }
-            
-            AssetDatabase.CreateAsset(comment, $"{folderPath}/{comment.name}.asset");
-
-            return comment;
         }
     }
 }