+++ Runtime/EditorCommentAsset.cs
@@ -1,13 +1,26 @@
-﻿using System;
+﻿#if UNITY_EDITOR
+using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Text;
 using Sirenix.OdinInspector;
+using UnityEditor;
 using UnityEngine;
 namespace OcUtility
 {
     [CreateAssetMenu(fileName = "New Editor Comment", menuName = "Utility/Editor Comment", order = 0)]
     public class EditorCommentAsset : ScriptableObject
     {
+        [DelayedProperty][ShowInInspector][PropertyOrder(-1)]
+        public string Name
+        {
+            get => name;
+            set
+            {
+                name = value;
+                AssetDatabase.RenameAsset(AssetDatabase.GetAssetPath(this), value);
+            }
+        }
         public Context[] Contexts;
 
         void Reset()
@@ -36,6 +49,22 @@ namespace OcUtility
             }
 
             GUIUtility.systemCopyBuffer = sb.ToString();
+        }
+
+        public static EditorCommentAsset CreateAsset(string folderPath, string fileName)
+        {
+            if(!AssetDatabase.IsValidFolder(folderPath))
+            {
+                Directory.CreateDirectory(folderPath);
+            }
+
+            var asset = CreateInstance<EditorCommentAsset>();
+            var path = $"{folderPath}/{fileName}.asset";
+            var uniquePath = AssetDatabase.GenerateUniqueAssetPath(path);
+            asset.name = Path.GetFileName(uniquePath);
+            AssetDatabase.CreateAsset(asset, uniquePath);
+
+            return asset;
         }
 
         [Serializable]
@@ -68,3 +97,4 @@ namespace OcUtility
         }
     }
 }
+#endif